fix(master): handle the problem that master nodes gains the same slot id
